fix: 메시지 전송 시 프로필 이미지 미반영 문제 수정
- 기존 toResponse에서 JWT 토큰의 profileImg를 사용하여
  프로필 업데이트 후 새로고침 전까지 구버전 이미지가 노출되는 문제 수정

- 로그인 시 Redis에 profileImage 적재
- 프로필 업데이트 시 Redis 캐시 갱신
- 로그아웃 시 Redis 캐시 삭제
- 메시지 전송/수정 시 Redis에서 최신 profileImage 조회하여 broadcast
