+++ backend/src/main/java/project/backend/domain/member/app/MemberService.java
@@ -3,7 +3,6 @@ package project.backend.domain.member.app;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -13,7 +12,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-import project.backend.domain.chat.chatroom.app.ChatRoomRedisService;
 import project.backend.domain.imagefile.ImageFileService;
 import project.backend.domain.member.dao.MemberRepository;
 import project.backend.domain.member.dto.MemberSearchResponse;
@@ -41,7 +39,7 @@ public class MemberService {
     private final ImageFileService imageFileService;
     private final PasswordEncoder passwordEncoder;
     private final ApplicationEventPublisher eventPublisher;
-    private final ChatRoomRedisService chatRoomRedisService;
+    private final MemberRedisService memberRedisService;
 
     @Value("${file.images.profile.default}")
     private String defaultProfileImg;
@@ -90,7 +88,7 @@ public class MemberService {
             String profileImage = imageFileService.saveProfileImage(file);
             targetMember.updateProfileImage(profileImage);
 
-            chatRoomRedisService.setProfileImage(targetMember.getId(), profileImage);
+            memberRedisService.setProfileImage(targetMember.getId(), profileImage);
         }
     }
